Update/requirements.txt
flask
python-dotenv
pymongo
dnspython
# pip install -r requirements.txt